test(cli): remove stale create helper; share staging helper; fix -d help text
